+++ include/json.h
@@ -1,7 +1,7 @@
 #ifndef JSON_H
 #define JSON_H
 
-#include "../on.h"
+#include "on.h"
 
 char *json_dumps(on *o);
 int json_dump(on *o, const char* filename);
